requirements.txt

pandas
matplotlib
google-api-python-client
textblob 
python-dotenv 
nltk
streamlit 